[♻️refactor]: store 위치 변경
- src/store 디렉토리에 user 폴더 추가하여
  로그인, 회원 정보, 토큰 관련 store파일 이동
