+++ src/components/Pages/Home/MainContent.tsx
@@ -4,10 +4,10 @@ import { OnboardingBanner, ImageContainer } from "@/components";
 import { useLoginMutation } from "@/hooks/api/useLoginMutation";
 import { useSocialLoginQuery } from "@/hooks/api/useSocialLoginQuery";
 import { useUserInfoQuery } from "@/hooks/api/useUserInfoQuery";
-import { useSignupUserStore } from "@/hooks/store/useSignupUserStore";
-import { useSocialLoginStore } from "@/hooks/store/useSocialLoginStore";
-import { useTokenStore } from "@/hooks/store/useTokenStore";
-import { useUserInfoStore } from "@/hooks/store/useUserInfoStore";
+import { useSignupUserStore } from "@/store/user/useSignupUserStore";
+import { useSocialLoginStore } from "@/store/user/useSocialLoginStore";
+import { useTokenStore } from "@/store/user/useTokenStore";
+import { useUserInfoStore } from "@/store/user/useUserInfoStore";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useEffect } from "react";
 
@@ -55,7 +55,7 @@ export default function MainContent() {
 
   const { data: userInfoData, isSuccess } = useUserInfoQuery(
     accessToken || "",
-    memberId
+    memberId,
   );
 
   useEffect(() => {
