+++ src/components/Pages/Profile/Edit/ProfileEditProfile/ProfileEditProfileNickname.tsx
@@ -2,8 +2,8 @@ import { Button, InputField } from "@/components";
 import { ButtonStyle } from "@/components/Button/Button.types";
 import { useState } from "react";
 import validateNickname from "@/utils/validateNickname";
-import { useProfileEditStore } from "@/hooks/store/useProfileEditStore";
-import { useUserInfoStore } from "@/hooks/store/useUserInfoStore";
+import { useProfileEditStore } from "@/store/user/useProfileEditStore";
+import { useUserInfoStore } from "@/store/user/useUserInfoStore";
 import * as S from "./ProfileEditProfileNickname.style";
 
 export default function ProfileEditProfileNickname() {
