+++ src/hooks/api/useSignupMutation.ts
@@ -2,7 +2,7 @@ import { postSignup } from "@/api/signup";
 import { ISignupUser } from "@/types/api/user";
 import { useMutation } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
-import { useTokenStore } from "../store/useTokenStore";
+import { useTokenStore } from "../../store/user/useTokenStore";
 
 // eslint-disable-next-line import/prefer-default-export
 export function useSignupMutation() {
